Fix Modular build: resolve circular includes, shared functions, and anonymous namespace consolidation
- Extract StringLiteralTokenUtils inline functions to .cpp to break circular include
- Fix TemplateEnvironmentSnapshotNode pointer vs reference mismatches across all call sites
- Move shared template instantiation helpers (applyTemplateArgIndirection, propagateFunctionSignatureFromTemplateArg, rebuildOuterTemplateParameter) into ParserTemplateClassShared.h
- Consolidate three anonymous namespace blocks into one in ConstExprEvaluator_Core.cpp
- Add missing includes (<span>, Parser.h, ExpressionSubstitutor.h, etc.) for Modular TU isolation
- Register StringLiteralTokenUtils.cpp in vcxproj and CompilerIncludes.h unity shard
- Document Modular-only crash in KNOWN_ISSUES.md
