+++ FlashCppMSVC.vcxproj
@@ -392,6 +392,9 @@
     <ClCompile Include="src\SemanticAnalysis.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)'!='Modular'">true</ExcludedFromBuild>
     </ClCompile>
+    <ClCompile Include="src\StringLiteralTokenUtils.cpp">
+      <ExcludedFromBuild Condition="'$(Configuration)'!='Modular'">true</ExcludedFromBuild>
+    </ClCompile>
     <ClCompile Include="src\TypeTraitEvaluator.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)'!='Modular'">true</ExcludedFromBuild>
     </ClCompile>
@@ -470,6 +473,7 @@
     <ClInclude Include="src\ProfilingTimer.h" />
     <ClInclude Include="src\StackString.h" />
     <ClInclude Include="src\StringBuilder.h" />
+    <ClInclude Include="src\StringLiteralTokenUtils.h" />
     <ClInclude Include="src\StringTable.h" />
     <ClInclude Include="src\SemanticAnalysis.h" />
     <ClInclude Include="src\SymbolTable.h" />
