+++ src/AstNodeTypes_Template.h
@@ -1,4 +1,5 @@
 #pragma once
+struct TemplateEnvironmentSnapshotNode;
 #include <span>
 #include <type_traits>
 #include "AstNodeTypes_DeclNodes.h"
@@ -416,21 +417,15 @@ public:
 				outer_template_args_.push_back(toTemplateArgInfo(arg));
 			}
 		}
-		outer_template_environment_snapshot_ = buildTemplateEnvironmentSnapshot(
-			std::span<const StringHandle>(outer_template_param_names_.data(), outer_template_param_names_.size()),
-			std::span<const TypeInfo::TemplateArgInfo>(outer_template_args_.data(), outer_template_args_.size()));
+		outer_template_environment_snapshot_node_ = nullptr;
 	}
 
-	void set_outer_template_bindings(const TemplateEnvironmentSnapshot& snapshot) {
-		outer_template_environment_snapshot_ = snapshot;
-		populateTemplateEnvironmentLegacyViews(
-			outer_template_environment_snapshot_,
-			outer_template_param_names_,
-			outer_template_args_);
+	void set_outer_template_bindings(const TemplateEnvironmentSnapshotNode* snapshot_node) {
+		outer_template_environment_snapshot_node_ = snapshot_node;
 	}
 
-	bool has_outer_template_bindings() const { return hasTemplateEnvironmentSnapshotBindings(outer_template_environment_snapshot_); }
-	const TemplateEnvironmentSnapshot& outer_template_environment_snapshot() const { return outer_template_environment_snapshot_; }
+	bool has_outer_template_bindings() const { return outer_template_environment_snapshot_node_ != nullptr; }
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot() const { return outer_template_environment_snapshot_node_; }
 	const InlineVector<StringHandle, 4>& outer_template_param_names() const { return outer_template_param_names_; }
 	const InlineVector<TypeInfo::TemplateArgInfo, 4>& outer_template_args() const { return outer_template_args_; }
 
@@ -442,7 +437,7 @@ private:
 	bool is_thread_local_ = false;
 	bool is_constexpr_;
 	bool is_constinit_;
-	TemplateEnvironmentSnapshot outer_template_environment_snapshot_;
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot_node_{};
 	InlineVector<StringHandle, 4> outer_template_param_names_;
 	InlineVector<TypeInfo::TemplateArgInfo, 4> outer_template_args_;
 };
@@ -680,21 +675,15 @@ public:
 				outer_template_args_.push_back(toTemplateArgInfo(arg));
 			}
 		}
-		outer_template_environment_snapshot_ = buildTemplateEnvironmentSnapshot(
-			std::span<const StringHandle>(outer_template_param_names_.data(), outer_template_param_names_.size()),
-			std::span<const TypeInfo::TemplateArgInfo>(outer_template_args_.data(), outer_template_args_.size()));
+		outer_template_environment_snapshot_node_ = nullptr;
 	}
 
-	void set_outer_template_bindings(const TemplateEnvironmentSnapshot& snapshot) {
-		outer_template_environment_snapshot_ = snapshot;
-		populateTemplateEnvironmentLegacyViews(
-			outer_template_environment_snapshot_,
-			outer_template_param_names_,
-			outer_template_args_);
+	void set_outer_template_bindings(const TemplateEnvironmentSnapshotNode* snapshot_node) {
+		outer_template_environment_snapshot_node_ = snapshot_node;
 	}
 
-	bool has_outer_template_bindings() const { return hasTemplateEnvironmentSnapshotBindings(outer_template_environment_snapshot_); }
-	const TemplateEnvironmentSnapshot& outer_template_environment_snapshot() const { return outer_template_environment_snapshot_; }
+	bool has_outer_template_bindings() const { return outer_template_environment_snapshot_node_ != nullptr; }
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot() const { return outer_template_environment_snapshot_node_; }
 	const InlineVector<StringHandle, 4>& outer_template_param_names() const { return outer_template_param_names_; }
 	const InlineVector<TypeInfo::TemplateArgInfo, 4>& outer_template_args() const { return outer_template_args_; }
 	void set_lazy_member_registry_key(StringHandle key) { lazy_member_registry_key_ = key; }
@@ -721,7 +710,7 @@ private:
 	bool has_template_initializer_list_ = false;
 	SaveHandle template_body_position_handle_;  // Handle to saved position for template body
 	SaveHandle template_initializer_list_position_handle_{};  // Handle to saved position for constructor initializer list
-	TemplateEnvironmentSnapshot outer_template_environment_snapshot_;
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot_node_{};
 	InlineVector<StringHandle, 4> outer_template_param_names_;
 	InlineVector<TypeInfo::TemplateArgInfo, 4> outer_template_args_;
 	TypeIndex owning_type_index_{};
@@ -847,21 +836,15 @@ public:
 				outer_template_args_.push_back(toTemplateArgInfo(arg));
 			}
 		}
-		outer_template_environment_snapshot_ = buildTemplateEnvironmentSnapshot(
-			std::span<const StringHandle>(outer_template_param_names_.data(), outer_template_param_names_.size()),
-			std::span<const TypeInfo::TemplateArgInfo>(outer_template_args_.data(), outer_template_args_.size()));
+		outer_template_environment_snapshot_node_ = nullptr;
 	}
 
-	void set_outer_template_bindings(const TemplateEnvironmentSnapshot& snapshot) {
-		outer_template_environment_snapshot_ = snapshot;
-		populateTemplateEnvironmentLegacyViews(
-			outer_template_environment_snapshot_,
-			outer_template_param_names_,
-			outer_template_args_);
+	void set_outer_template_bindings(const TemplateEnvironmentSnapshotNode* snapshot_node) {
+		outer_template_environment_snapshot_node_ = snapshot_node;
 	}
 
-	bool has_outer_template_bindings() const { return hasTemplateEnvironmentSnapshotBindings(outer_template_environment_snapshot_); }
-	const TemplateEnvironmentSnapshot& outer_template_environment_snapshot() const { return outer_template_environment_snapshot_; }
+	bool has_outer_template_bindings() const { return outer_template_environment_snapshot_node_ != nullptr; }
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot() const { return outer_template_environment_snapshot_node_; }
 	const InlineVector<StringHandle, 4>& outer_template_param_names() const { return outer_template_param_names_; }
 	const InlineVector<TypeInfo::TemplateArgInfo, 4>& outer_template_args() const { return outer_template_args_; }
 	void set_lazy_member_registry_key(StringHandle key) { lazy_member_registry_key_ = key; }
@@ -877,7 +860,7 @@ private:
 	bool has_noexcept_specifier_ = false;  // True iff an explicit noexcept / noexcept(expr) was written
 	bool is_constexpr_ = false;  // True iff the destructor was declared with 'constexpr'
 	std::optional<ASTNode> noexcept_expression_;	 // For explicit noexcept(expr)
-	TemplateEnvironmentSnapshot outer_template_environment_snapshot_;
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot_node_{};
 	InlineVector<StringHandle, 4> outer_template_param_names_;
 	InlineVector<TypeInfo::TemplateArgInfo, 4> outer_template_args_;
 	BodyStateTag body_state_tag_ = BodyStateTag::NotMaterialized;
@@ -1413,21 +1396,15 @@ public:
 				outer_template_args_.push_back(toTemplateArgInfo(arg));
 			}
 		}
-		outer_template_environment_snapshot_ = buildTemplateEnvironmentSnapshot(
-			std::span<const StringHandle>(outer_template_param_names_.data(), outer_template_param_names_.size()),
-			std::span<const TypeInfo::TemplateArgInfo>(outer_template_args_.data(), outer_template_args_.size()));
+		outer_template_environment_snapshot_node_ = nullptr;
 	}
 
-	void set_outer_template_bindings(const TemplateEnvironmentSnapshot& snapshot) {
-		outer_template_environment_snapshot_ = snapshot;
-		populateTemplateEnvironmentLegacyViews(
-			outer_template_environment_snapshot_,
-			outer_template_param_names_,
-			outer_template_args_);
+	void set_outer_template_bindings(const TemplateEnvironmentSnapshotNode* snapshot_node) {
+		outer_template_environment_snapshot_node_ = snapshot_node;
 	}
 
-	bool has_outer_template_bindings() const { return hasTemplateEnvironmentSnapshotBindings(outer_template_environment_snapshot_); }
-	const TemplateEnvironmentSnapshot& outer_template_environment_snapshot() const { return outer_template_environment_snapshot_; }
+	bool has_outer_template_bindings() const { return outer_template_environment_snapshot_node_ != nullptr; }
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot() const { return outer_template_environment_snapshot_node_; }
 	const InlineVector<StringHandle, 4>& outer_template_param_names() const { return outer_template_param_names_; }
 	const InlineVector<TypeInfo::TemplateArgInfo, 4>& outer_template_args() const { return outer_template_args_; }
 
@@ -1452,7 +1429,7 @@ private:
 	bool has_deleted_copy_constructor_ = false;		// Track deleted copy constructor
 	bool has_deleted_move_constructor_ = false;		// Track deleted move constructor
 	std::vector<DeferredStaticAssert> deferred_static_asserts_;	// Static_asserts deferred during template definition
-	TemplateEnvironmentSnapshot outer_template_environment_snapshot_;
+	const TemplateEnvironmentSnapshotNode* outer_template_environment_snapshot_node_{};
 	InlineVector<StringHandle, 4> outer_template_param_names_;
 	InlineVector<TypeInfo::TemplateArgInfo, 4> outer_template_args_;
 	StructBodyStateTag struct_body_state_tag_ = StructBodyStateTag::NotMaterialized;
