+++ src/Parser_Templates_Inst_MemberFunc.cpp
@@ -1,5 +1,6 @@
 #include "Parser.h"
 #include "ConstExprEvaluator.h"
+#include "ExpressionSubstitutor.h"
 #include "NameMangling.h"
 #include "OverloadResolution.h"
 #include "ParserTemplateClassShared.h"
@@ -1195,17 +1196,14 @@ std::optional<ASTNode> Parser::try_instantiate_constructor_template(
 	InlineVector<StringHandle, 4> outer_param_names;
 	InlineVector<TypeInfo::TemplateArgInfo, 4> outer_args;
 	populateTemplateEnvironmentLegacyViews(
-		materialization_source_ctor->outer_template_environment_snapshot(),
+		TemplateEnvironmentSnapshot{materialization_source_ctor->outer_template_environment_snapshot()},
 		outer_param_names,
 		outer_args);
-	const TemplateEnvironmentSnapshot* outer_parent_snapshot =
-		materialization_source_ctor->has_outer_template_bindings()
-			? &materialization_source_ctor->outer_template_environment_snapshot()
-			: nullptr;
+	TemplateEnvironmentSnapshot outer_parent_snapshot{materialization_source_ctor->outer_template_environment_snapshot()};
 	lazy_info.outer_template_environment_snapshot = buildTemplateEnvironmentSnapshotFromBindings(
 		template_params,
 		ctor_template_args,
-		outer_parent_snapshot);
+		&outer_parent_snapshot);
 	for (StringHandle outer_name : outer_param_names) {
 		Token outer_token(Token::Type::Identifier, StringTable::getStringView(outer_name), 0, 0, 0);
 		lazy_info.template_params.push_back(TemplateParameterNode(outer_name, outer_token));
